improve config
