+++ .babelrc.js
@@ -5,7 +5,7 @@ module.exports = {
 		'@babel/plugin-proposal-export-default-from',
 		'@babel/plugin-transform-modules-commonjs',
 		'@babel/plugin-proposal-object-rest-spread',
-		'@babel/plugin-transform-runtime',
+		// ['@babel/plugin-transform-runtime'],
 		'@babel/plugin-proposal-function-bind',
 		['@babel/plugin-proposal-decorators', { legacy: true }],
 		['@babel/plugin-proposal-class-properties', { loose: true }],
@@ -19,21 +19,9 @@ module.exports = {
 	],
 	presets: [
 		[
-			'@babel/preset-env',
-			{
-				"targets": {
-					"chrome": "49",
-					"ie": "10"
-				},
-				corejs: 3,
-				modules: false,
-				useBuiltIns: 'entry',
-				exclude: ['transform-typeof-symbol'],
-			},
+			'@babel/preset-env'
 		],
-		['@babel/preset-react', {
-			useBuiltIns: true,
-		}],
+		['@babel/preset-react'],
 		['@babel/preset-typescript'],
 	],
 };
