+++ src/App/actions.ts
@@ -1,0 +1,5 @@
+const update = () => ({name: 'new app'});
+
+export default {
+	update,
+}
