+++ src/App/index.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import { Route, Link } from 'react-router-dom';
 import TransitionSwitch from '../router/TransitionSwitch';
+import Inject from '../store/inject';
 import routes from '../router';
 import './style.scss';
 
 const App: React.FunctionComponent = (p: any) => {
+	console.log(p);
 	return (
 		<>
 			<Link to="/page1">page1</Link>
@@ -17,4 +19,4 @@ const App: React.FunctionComponent = (p: any) => {
 	);
 };
 
-export default App;
+export default Inject('app')(App);
