+++ src/App/state.ts
@@ -1,0 +1,3 @@
+export default {
+	name: 'app',
+}
