+++ src/Charts/Map/style.scss
@@ -1,4 +1,4 @@
-.page3 {
+.page2 {
 	width: 100%;
 	height: 100vh;
 	background-color: lightgreen;
