+++ src/index.tsx
@@ -1,10 +1,9 @@
-// import 'babel-polyfill';
+import 'babel-polyfill';
 import React from 'react';
 import ReactDom from 'react-dom';
 import { HashRouter } from 'react-router-dom';
 import {store, Provider} from './store';
 import App from './App'
-
 
 const content = (
 	<Provider store={store}>
