+++ src/pages/Page2/view.tsx
@@ -1,8 +1,8 @@
 import React, {useEffect} from 'react';
 import style from './style.scss';
 import Inject from '@client/store/inject'
-import Pack from '@client/Charts/Pack'
-import data from '@client/Charts/Pack/flare.json'
+// import Pack from '@client/Charts/Pack'
+// import data from '@client/Charts/Pack/flare.json'
 
 const Page2: React.FC<any> = (p: any) => {
 	const { page1List: {state, actions} } = p;
