+++ src/store/index.ts
@@ -1,7 +1,9 @@
 import createStore from './createStore';
 import Provider from './Provider';
+import appState from '../App/state';
+import appActions from '../App/actions';
 
-const store = createStore();
+const store = createStore({appActions}, {appState});
 
 export {
 	store,
